+++ tests/AddressMiscGettersTest.php
@@ -61,11 +61,23 @@ class AddressMiscGettersTest extends TestCase
         self::assertFalse($ipNotV4->isCompatibleIPv4());
 
         self::assertTrue($ipMapped->isIPv4());
-        self::assertTrue($ipCompat->isIPv4());
+        self::assertFalse($ipCompat->isIPv4());
         self::assertFalse($ipNotV4->isIPv4());
 
         self::assertEquals('64.92.175.4', (string)$ipMapped->getIPv4());
-        self::assertEquals('64.92.175.4', (string)$ipCompat->getIPv4());
+    }
+
+    public function testLocalhostIsNotIPv4(): void
+    {
+        $localhost = IPv6Address::fromString('::1');
+
+        self::assertFalse($localhost->isIPv4());
+        self::assertTrue($localhost->isCompatibleIPv4()); // this is why compatible range is problematic
+
+        $this->expectException(DomainException::class);
+        $this->expectExceptionMessage('This IPv6 address does not encode IPv4');
+
+        $localhost->getIPv4();
     }
 
     public function testIPv4NotEncodedInIPv6(): void